third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/white-space/seg-break-transformation-008.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] linebreak only
  assert_equals: expected 195 but got 201
[FAIL] spaces linebreak
  assert_equals: expected 195 but got 201
[FAIL] linebreak spaces
  assert_equals: expected 195 but got 201
[FAIL] spaces linebreak spaces
  assert_equals: expected 195 but got 201
[FAIL] multiple linebreaks
  assert_equals: expected 195 but got 201
[FAIL] multiple linebreaks + spaces
  assert_equals: expected 195 but got 201
Harness: the test ran to completion.
